预售小程序显示处理

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/wxapp/order/OrderListMpVo.java
@@ -83,8 +83,8 @@ public class OrderListMpVo{
 	private List<String> orderType;
     @JsonIgnore
 	private Byte orderPayWay;
-    /**补款*/
-    private BigDecimal bkReturnType;
+    /**定金是否退款*/
+    private Byte bkReturnType;
     /**定金抵扣金额*/
     private BigDecimal preDiscountMoney;
     /**预售订单预计发货时间*/
